==> Version_ENG/07_Consolidated_Person_Number.qmd <==

```{r setup, include=FALSE, message=FALSE, error=FALSE, warning=FALSE}
knitr::opts_chunk$set(
  echo = TRUE,
  message = FALSE,
  warning = FALSE,
  cache = TRUE)
library(kableExtra)
library(tidyverse)
library(magrittr)
rstan::rstan_options(auto_write = TRUE) # speed up running time 
tba <- function(dat, cap = NA){
  kable(dat,
      format = "html", digits =  4,
      caption = cap) %>% 
     kable_styling(bootstrap_options = "striped", full_width = F)%>%
         kable_classic(full_width = F, html_font = "Arial Narrow")
}
```

# consolidating the number of people per household

```{r, eval=FALSE}
### Cleaning R environment ###
rm(list = ls())

#################
### Libraries ###
#################
library(tidyverse)
library(data.table)
library(openxlsx)
library(magrittr)
library(lme4)  # For fitting linear mixed-effects models
library(rstan)  # For Bayesian data analysis using Stan
library(rstanarm)  # For fitting Bayesian regression models
library(merTools)
cat("\f")
```

-   Read census data and models. 

```{r, eval=FALSE}
censo_vivienda <-
  readRDS("Recursos/05_Model_for_people/Data/01_censo_vivienda_desocupadas.rds")
Base_ugms <- readRDS("Recursos/05_Model_for_people/Data/Base_ugms_estandarizada.rds")

modelo_todas <-
  readRDS("Recursos/05_Model_for_people/Data/fit_poisson_todas.rds")
modelo_ocupadas <-
  readRDS("Recursos/05_Model_for_people/Data/fit_poisson_ocupadas.rds")
```

-   Make predictions using the models

```{r, eval=FALSE}
pred_todas <- predict(modelo_todas,
                      newdata =  Base_ugms,
                      type = "response",
                      se.fit = TRUE)

saveRDS(pred_todas,file = "Recursos/05_Model_for_people/Data/pred_todas.rds")

pred_todas <-  readRDS(file = "Recursos/05_Model_for_people/Data/pred_todas.rds")

Base_ugms$pred_todas <- pred_todas$fit 
Base_ugms$pred_todas_se <- pred_todas$se.fit

hist(Base_ugms$pred_todas)

pred_ocupadas <- predict(modelo_ocupadas,
                         newdata =  Base_ugms,
                         type = "response",
                         se.fit = TRUE)

saveRDS(pred_ocupadas,file = "Recursos/05_Model_for_people/Data/pred_ocupadas.rds")

pred_ocupadas <-  readRDS(file = "Recursos/05_Model_for_people/Data/pred_ocupadas.rds")

Base_ugms$pred_ocupadas <- pred_ocupadas$fit
Base_ugms$pred_ocupadas_se <- pred_ocupadas$se.fit

hist(Base_ugms$pred_ocupadas)

summary(Base_ugms$pred_todas)
summary(Base_ugms$pred_ocupadas)
```

-   Merge census information with predictions

```{r, eval=FALSE}
censo_temp <- censo_vivienda %>%
  full_join(Base_ugms) %>%
  group_by(UGM_ID) %>%
  mutate(
    pred_conteos = case_when(
      greenpoint2  == "Sin informacion pero  n>0" ~ pred_ocupadas,
      greenpoint2  == "Sin informacion pero n>=0"  ~ pred_todas ,
      TRUE ~ H01A_TOTAL_PERSONAS
    ),
    MEInf_pred_conteos = case_when(
      greenpoint2  == "Sin informacion pero  n>0" ~ 1.96*pred_ocupadas_se,
      greenpoint2  == "Sin informacion pero n>=0"  ~ 1.96*pred_todas_se ,
      TRUE ~ 0
    ),
    MESup_pred_conteos = case_when(
      greenpoint2  == "Sin informacion pero  n>0" ~ 1.96*pred_ocupadas_se,
      greenpoint2  == "Sin informacion pero n>=0"  ~ 1.96*pred_todas_se ,
      TRUE ~ 0
    )
  )
```

-   Calculate sum of predictions

```{r, eval=FALSE}
sum(censo_temp$pred_conteos)

# Summary of estimates per UGM
censo_temp %>%  
  group_by(UGM_ID) %>% 
  summarise(Min_ugm = min(pred_conteos),
            Max_ugm = max(pred_conteos)) %>% 
  View()

# Summary of estimates filtered by greenpoint2
censo_temp %>% filter(greenpoint2  == "Sin informacion pero n>=0" ) %>% 
  group_by(UGM_ID) %>% 
  summarise(Min_ugm = min(pred_conteos),
            Max_ugm = max(pred_conteos))%>% 
  View()
```

-   Save consolidated information to a file

```{r, eval=FALSE}
saveRDS(censo_temp, 
        file = "Recursos/05_Model_for_people/Data/02_censo_vivienda_personas.rds")
```
